DasharoPayloadPkg/Library/DxeCapsuleLibFmp: Implement capsule authentication
This Commit Implements capsule authentication in DxeCapsuleLibFmp. It
uses an FMP capsule inside another capsule (FMP payload as a capsule).
The top caspule should only contain one payload and nothing else.
This way, when the payload is verified, we have the entire bottom capsule
authenticated. This is mainly required for the drivers loading. After
that the drivers are responsible to authenticate the payloads of the
inner capsule again.

Signed-off-by: Gustavo dos Santos Cardoso <gustavo_16a@hotmail.com>